Move placeholder logic into separate plugin
